# swagger.yaml
openapi: 3.0.0
info:
  title: API Quản lý Nông Trại Nấm
  version: 1.0.0
  description: Tài liệu API cho Hệ thống Quản lý Nuôi trồng Nấm
servers:
  - url: /
    description: Máy chủ đang phục vụ Swagger (tự dùng host và cổng hiện tại)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Mushroom:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        scientificName:
          type: string
        commonName:
          type: string
        otherNames:
          type: string
        family:
          type: string
        genus:
          type: string
        capDescription:
          type: string
        gillsDescription:
          type: string
        stemDescription:
          type: string
        sporePrintColor:
          type: string
        bruisingBehavior:
          type: string
        ecologyType:
          type: string
          enum: [SAPROBIC, MYCORRHIZAL, PARASITIC]
        habitat:
          type: string
        fruitingSeason:
          type: string
        edibilityStatus:
          type: string
          enum: [CHOICE, EDIBLE, INEDIBLE, POISONOUS, DEADLY]
        toxicitySymptoms:
          type: string
        medicinalProperties:
          type: string
        cultivationDifficulty:
          type: string
          enum: [EASY, MEDIUM, HARD, UNCULTIVABLE]
        imageUrl:
          type: string
        coverImageUrl:
          type: "string"
          nullable: true
          readOnly: true
        imageCount:
          type: "integer"
          readOnly: true
...
security:
  - bearerAuth: []
paths:
...
  /api/mushroom-species:
    get:
      summary: Lấy danh sách giống nấm
      tags: [Quản lý Giống Nấm]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Tìm kiếm theo tên khoa học, tên thường gọi, họ, chi
      responses:
        '200':
          description: Danh sách các giống nấm (kèm phân trang)
    post:
      summary: Thêm giống nấm mới
      tags: [Quản lý Giống Nấm]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mushroom'
      responses:
        '201':
          description: Thêm giống nấm thành công
  /api/mushroom-species/{id}:
    get:
      summary: Chi tiết giống nấm và thông tin gallery
      tags: [Quản lý giống nấm]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Chi tiết giống nấm
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Mushroom'
        '401': { description: Token không hợp lệ }
        '403': { description: Không đủ quyền }
        '404': { description: Không tìm thấy giống nấm }
    put:
      summary: Cập nhật thông tin giống nấm
      tags: [Quản lý Giống Nấm]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID của giống nấm
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Mushroom'
      responses:
        '200':
          description: Cập nhật thông tin thành công
    delete:
      summary: Xóa một giống nấm
      tags: [Quản lý Giống Nấm]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID của giống nấm cần xóa
      responses:
        '200':
          description: Xóa giống nấm thành công
...